+++ src/sentry_options.h
@@ -48,7 +48,7 @@ struct sentry_options_s {
     bool cache_keep;
 
     int cache_max_age; // TODO in days?
-    int cache_max_size; // TODO in kb?
+    size_t cache_max_size; // TODO in kb?
 
     sentry_attachment_t *attachments;
     sentry_run_t *run;
